TST: method-specific files for DataFrame assign, interpolate  (#32110)
